requirements.txt
torch==2.9.1
transformers==4.57.3
trl==0.26.2
datasets==4.4.2
peft==0.7.1
bitsandbytes==0.49.0
accelerate==1.12.0
scikit-learn==1.7.2
numpy==1.26.4
pandas==2.2.3
matplotlib==3.9.2
pyyaml==6.0.3
wandb==0.23.1
sentencepiece==0.2.1
safetensors==0.7.0
requests==2.32.3
python-dotenv==1.0.1
tiktoken==0.11.0
